Refactor to handle total price logic in useCart hook
CartList component is refactored to remove code that was handling the total prices sum to just render the total prop passed, the total logic is now handled in the useCart custom hook.
CartDropdown component is refactored to receive total as a string instead of number.
useCart hook adds the logic and types to handle total price sum, also adds tests to check for the correct behaviour of the hook.