Add test-utils.tsx to render providers in tests
Add customRender function that renders the ThemeProvider and CartContext.Provider around the element passed. The customRender function is exported as render to be used in the tests.
Remove tests/helpers.tsx util that was being used to render the ThemeProvider aound the element passed, this is now being handled in the new test-utils.tsx.